+++ packages/analytics-js/src/components/configManager/util/commonUtil.ts
@@ -24,6 +24,8 @@ import { clone } from 'ramda';
 import type { PluginName } from '@rudderstack/analytics-js-common/types/PluginsManager';
 import { isValidURL, removeDuplicateSlashes } from '@rudderstack/analytics-js-common/utilities/url';
 import { removeLeadingPeriod } from '@rudderstack/analytics-js-common/utilities/string';
+import type { Destination } from '@rudderstack/analytics-js-common/types/Destination';
+import { getDestinationUserFriendlyId } from '@rudderstack/analytics-js-common/utilities/destinations';
 import { MODULE_TYPE, APP_VERSION } from '../../../constants/app';
 import { BUILD_TYPE, DEFAULT_CONFIG_BE_URL } from '../../../constants/urls';
 import { state } from '../../../state';
@@ -44,13 +46,13 @@ import {
 import { getDomain, removeTrailingSlashes } from '../../utilities/url';
 import type { ConfigResponseDestinationItem, SourceConfigResponse } from '../types';
 import {
+  CUSTOM_DEVICE_MODE_DESTINATION_DISPLAY_NAME,
   DEFAULT_DATA_SERVICE_ENDPOINT,
   DEFAULT_STORAGE_ENCRYPTION_VERSION,
   StorageEncryptionVersionsToPluginNameMap,
 } from '../constants';
 import { getDataServiceUrl, isValidStorageType, isWebpageTopLevelDomain } from './validate';
 import { getConsentManagementData } from '../../utilities/consent';
-import type { Destination } from '@rudderstack/analytics-js-common/types/Destination';
 
 /**
  * Determines the SDK URL
@@ -405,9 +407,12 @@ const getDestinationsFromConfig = (destinations: ConfigResponseDestinationItem[]
     config: destination.config,
     shouldApplyDeviceModeTransformation: destination.shouldApplyDeviceModeTransformation ?? false,
     propagateEventsUntransformedOnError: destination.propagateEventsUntransformedOnError ?? false,
-    userFriendlyId: `${destination.destinationDefinition.displayName.replaceAll(' ', '-')}___${
-      destination.id
-    }`,
+    userFriendlyId: getDestinationUserFriendlyId(
+      destination.destinationDefinition.displayName,
+      destination.id,
+    ),
+    isCustomIntegration:
+      destination.destinationDefinition.displayName === CUSTOM_DEVICE_MODE_DESTINATION_DISPLAY_NAME,
   }));
 
 export {