+++ packages/analytics-js-plugins/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.12.0](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-plugins@3.11.1...@rudderstack/analytics-js-plugins@3.12.0) (2025-08-11)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.23.0`
+* `@rudderstack/analytics-js-cookies` updated to version `0.5.1`
+* `@rudderstack/analytics-js-integrations` updated to version `3.16.1`
+
+### Features
+
+* add custom device mode integrations support ([#2309](https://github.com/rudderlabs/rudder-sdk-js/issues/2309)) ([db078e6](https://github.com/rudderlabs/rudder-sdk-js/commit/db078e6bae9ab57a18003ac3fa231be8d94cdcaa)), closes [#2295](https://github.com/rudderlabs/rudder-sdk-js/issues/2295) [#2299](https://github.com/rudderlabs/rudder-sdk-js/issues/2299)
+
 ## [3.11.1](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-plugins@3.11.0...@rudderstack/analytics-js-plugins@3.11.1) (2025-07-30)
 
 ### Dependency Updates