+++ packages/analytics-js-service-worker/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.3.1](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-service-worker@3.3.0...@rudderstack/analytics-js-service-worker@3.3.1) (2025-08-11)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.23.0`
 ## [3.3.0](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-service-worker@3.2.27...@rudderstack/analytics-js-service-worker@3.3.0) (2025-07-21)
 
 ### Dependency Updates